% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimatic-predictions.R
\name{bioclim_by_year}
\alias{bioclim_by_year}
\title{Biennial bioclimatic estimates}
\usage{
bioclim_by_year(yr, model_for_prior_years, pred_grid)
}
\arguments{
\item{yr}{Year for which bioclimatic predictions will be estimated}

\item{model_for_prior_years}{sdmTMB model for all years prior to yr above}

\item{pred_grid}{Appropriate prediction grid containing unique covariate values 
for each year in model and the year to be estimated. Can include additional 
years before and/or after above period.}
}
\description{
Biennial bioclimatic estimates
}
