% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioclimatic-predictions.R
\name{bioclim_multi_year}
\alias{bioclim_multi_year}
\title{Bioclimatic estimates for multiyear period}
\usage{
bioclim_multi_year(yr, model_for_prior_years, pred_grid)
}
\arguments{
\item{yr}{First year for which bioclimatic predictions will be estimated}
}
\description{
Bioclimatic estimates for multiyear period
}
