% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocc.R
\name{dist_based_vocc}
\alias{dist_based_vocc}
\title{Velocity of climate change based on distance}
\usage{
dist_based_vocc(start_data, end_data, x = "x", y = "y",
  variable_names = c("var.1", "var.2"), plus_minus = c(0.5, 0.5),
  round_fact = NULL, min_thresholds = NULL, max_thresholds = NULL,
  match_logic = NULL, cell_size = 2, max_dist = NULL, delta_t = 50,
  raster = FALSE, kNN = FALSE)
}
\arguments{
\item{start_data}{Named list containing starting climate data.}

\item{end_data}{Named list containing target climate data.}

\item{x}{Numeric vector of x coordinates.}

\item{y}{Numeric vector of y coordinates.}

\item{variable_names}{Vector of columns/layers names within each data element.}

\item{plus_minus}{Vector of plus/minus threshold(s) to define a symetical climate match.}

\item{min_thresholds}{Optional vector of negative thresholds.
Include if sensitivity to the direction of climate change is not symmetrical and `match_logic` is NULL.}

\item{max_thresholds}{Optional vector of positive thresholds.
Include if sensitivity to the direction of climate change is not symmetrical and `match_logic` is NULL.}

\item{match_logic}{An optional vector of logical functions applied to rounded climate values.
If max_thresholds are not provided, the default of 'NULL' will apply symmetrical plus/minus thresholds to raw climate values.}

\item{cell_size}{Cell size in raster or distance to nearest data point.}

\item{max_dist}{Distance reported when no analogue found. Default will
estimate this assuming that both major axes are likely to hold an analogue.}

\item{delta_t}{Time difference between starting and target time periods.}

\item{raster}{Logical for whether climate data is in raster form.}

\item{kNN}{Logical for whether to use kNN search for extra speed
(but uses rounding, symmetrical thresholds, and only returns 1 target cell per source).}
}
\description{
Velocity of climate change based on distance
}
