% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mountains.R
\name{fixed_density}
\alias{fixed_density}
\title{Calculate density curve for non-time-varying, scaled, quadratic parameters}
\usage{
fixed_density(m, predictor = "temp")
}
\arguments{
\item{m}{Output of sdmTMB model}

\item{predictor}{Prefix for scaled parameter in model}
}
\description{
Calculate density curve for non-time-varying, scaled, quadratic parameters
}
