% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mountains.R
\name{fixed_interaction_density}
\alias{fixed_interaction_density}
\title{Calculate density curve for scaled, quadratic parameters with non-AR1 year interaction}
\usage{
fixed_interaction_density(m, predictor = "temp")
}
\arguments{
\item{m}{Output of sdmTMB model}

\item{predictor}{Prefix for scaled parameter in model}
}
\description{
Calculate density curve for scaled, quadratic parameters with non-AR1 year interaction
}
