% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vocc-interaction.R
\name{plot_chopstick_slopes}
\alias{plot_chopstick_slopes}
\title{Plot slopes from chopstick plots}
\usage{
plot_chopstick_slopes(
  slopedat,
  type = NULL,
  x_variable = "temp_trend_scaled",
  legend_position = c(0.7, 0.95),
  hack = F,
  imm_slopes = NULL,
  colours = NULL
)
}
\arguments{
\item{slopedat}{Df from chopstick_slopes function}

\item{type}{Select type if more than one interaction in model}

\item{x_variable}{}

\item{legend_position}{Given legend position}

\item{hack}{Logical for comparing with hacked slope estimate w interaction SE}

\item{imm_slopes}{Add immature data}

\item{colours}{Add custom colours}
}
\description{
Plot slopes from chopstick plots
}
