% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vocc-interaction.R
\name{plot_interaction}
\alias{plot_interaction}
\title{Plot interaction for vocc regression models}
\usage{
plot_interaction(model, species = NULL,
  variables = c("temp_vel_squashed", "do_vel_squashed"),
  choose_x = NULL, scaled = TRUE)
}
\arguments{
\item{model}{Model from vocc_regression function.
Requires named data element and coefs table.}

\item{species}{Species effect to be plotted.}

\item{variables}{List which variables are interacting.}

\item{scaled}{Logical for whether variables are scaled.}
}
\description{
Plot interaction for vocc regression models
}
