% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-mountains.R
\name{plot_mountains}
\alias{plot_mountains}
\title{Predicted curves for parameters in sdmTMB models}
\usage{
plot_mountains(dat, time_varying = TRUE, variable_label = "Depth",
  xlimits = c(0, max(dat$x)), x_breaks = waiver())
}
\arguments{
\item{dat}{Dataframe of density estimates containing x, y_hat, and year}

\item{time_varying}{Logical for if the parameter of interest is time-varying}

\item{variable_label}{Descriptive label for parameter "x"}

\item{xlimits}{X-axis limits. Default is c(0, max(dat$x)).}

\item{x_breaks}{Give breaks. Default is the ggplot default.}
}
\description{
Predicted curves for parameters in sdmTMB models
}
