% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_vocc}
\alias{plot_vocc}
\title{Plot climate change vectors}
\usage{
plot_vocc(df, fill_col = NULL, fill_label = NULL, raster_alpha = 1,
  vec_aes = "distance", vec_lwd_range = c(0.7, 0.8), vec_alpha = 1,
  max_vec_plotted = 100, low_fill = "Steel Blue 4",
  mid_fill = "white", high_fill = "Red 3", low_col = "white",
  mid_col = "white", high_col = "white", coast = NULL,
  contours = NULL)
}
\arguments{
\item{df}{Dataframe produced by dist_based_vocc function.}

\item{fill_col}{Vector name for colouring raster grid cells.}

\item{fill_label}{Label for legend for raster grid cell colour.}

\item{raster_alpha}{Raster transparency.}

\item{vec_aes}{Vector name for any plotting variation.}

\item{vec_lwd_range}{Range in vector line widths.}

\item{vec_alpha}{Vector transparency.}

\item{max_vec_plotted}{Upper limit for vector lenths. 
Cells with no match within this distance will be labeled 'NA'.}

\item{low_fill}{Colour of negative values if raster values span zero.}

\item{mid_fill}{Colour of zero value raster cells.}

\item{high_fill}{Colour of positive values if raster values span zero.}

\item{low_col}{Low colour for vectors.}

\item{mid_col}{Middle colour for vectors.}

\item{high_col}{High colour for vectors.}

\item{coast}{Coast polygons where (x = "X", y = "Y", group = "PID"). 
If NULL, will attempt to create them for xy values in df.}

\item{contours}{Polygons of contour lines where (x = "X", y = "Y", group = "paste(PID, SID)").
If NULL, will attempt to create bathymetry layer for xy values in df using gfplot.}
}
\description{
Plot climate change vectors
}
