% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-data-processing.R
\name{scale_predictors}
\alias{scale_predictors}
\title{Scale survey predictors}
\usage{
scale_predictors(dat, predictors = c(quo(log_depth)))
}
\arguments{
\item{predictors}{List of variable names must by wrapped in 'quo()'.}
}
\description{
Scale survey predictors
}
