% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vocc-interaction.R
\name{slope_scatterplot}
\alias{slope_scatterplot}
\title{Scatterplot of slopes against traits}
\usage{
slope_scatterplot(
  slopes_w_traits,
  x,
  slope_var = "slope_est",
  col_group = "chopstick",
  point_size = 2,
  regression = F
)
}
\arguments{
\item{slopes_w_traits}{Df of slopes merged with df of species traits.}

\item{x}{Variable (species trait) on x-axis.}

\item{slope_var}{Varible containing slope estimate.}

\item{col_group}{Varible to colour by.}

\item{point_size}{}

\item{regression}{Logical for plotting simple linear regression line.}
}
\description{
Scatterplot of slopes against traits
}
