% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mature-biomass-est.R
\name{split_catch_maturity}
\alias{split_catch_maturity}
\title{Split catch densities by mass ratio of maturity classes}
\usage{
split_catch_maturity(survey_sets, fish, bath, survey = c("SYN HS",
  "SYN QCS"), years = NULL, cutoff_quantile = 0.9995, plot = FALSE)
}
\arguments{
\item{survey_sets}{Data from get_survey_sets.}

\item{fish}{Data from get_survey_samples.}

\item{survey}{List of survey abbreviations.}

\item{years}{List of years. Default 'NULL' includes all years.}

\item{cutoff_quantile}{Set max cutoff for mass modeled from lengths.}

\item{plot}{Logical for whether to produce plots 
(length-weight and length-at-maturity relationships).}
}
\description{
Split catch densities by mass ratio of maturity classes
}
\examples{
events <- readRDS("analysis/VOCC/data/event-data-pop-1n3.rds")
fish <- readRDS("analysis/VOCC/data/bio-data-pop-1n3.rds")

biomass <- split_catch_maturity (events, fish, bath
  survey = c("SYN HS","SYN QCS"), 
  years = NULL,
  cutoff_quantile = 0.9995,
  plot = TRUE
)

}
