% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-data-processing.R
\name{interp_survey_bathymetry}
\alias{interp_survey_bathymetry}
\alias{add_missing_depths}
\alias{scale_predictors}
\title{Interpolate survey bathymetry}
\usage{
interp_survey_bathymetry(dat, utm_zone = 9)

add_missing_depths(dat, survey, years, bath)

scale_predictors(dat, predictors = c(quo(log_depth)))
}
\arguments{
\item{bath}{Output from [interp_survey_bathymetry()].}

\item{predictors}{List of variable names must by wrapped in 'quo()'.}
}
\description{
Interpolate survey bathymetry

Add missing depths

Scale survey predictors
}
