% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocc-gradient.R
\name{vocc_gradient_calc}
\alias{vocc_gradient_calc}
\title{Velocity of climate change}
\usage{
vocc_gradient_calc(data, layer, scale_fac = 1, time_step = "year",
  indices = NULL, divisor = 10, latlon = FALSE,
  quantile_cutoff = 0.05)
}
\arguments{
\item{data}{Gridded data.}

\item{layer}{Variable to retain. Defaults to "est", for raw sdmTMB predictions.}

\item{scale_fac}{Controls how the original 2-km projection is aggregated
(e.g. a value of 5 means that the raster would be reprojected to 10km grid).}

\item{time_step}{Time variable.}

\item{latlon}{Default is TRUE. False for UTMs or other true distance grid.}

\item{quantile_cutoff}{Used for plotting to set min and max angles of vectors.
Defaults to 0.05.}

\item{grad_time_steps}{Define time periods.
If NULL will use spatial gradient in last time period.}
}
\description{
Velocity of climate change
}
