% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vocc-interaction.R
\name{interaction_df}
\alias{interaction_df}
\title{Make prediction dataframe for chopsticks}
\usage{
interaction_df(
  data,
  model_matrix,
  species = NULL,
  x_variable = "squashed_temp_vel",
  split_variable = "mean_temp",
  use_quantiles = TRUE
)
}
\arguments{
\item{data}{The data going into the TMB model.}

\item{model_matrix}{The input model matrix to the TMB model.}

\item{species}{If NULL, loops through all species in raw data.}

\item{x_variable}{Which variable to plot sticks on.}

\item{split_variable}{Which variable to plot sticks at the min and max of.}

\item{use_quantiles}{If TRUE, plots lines for 25th and 75th quantile of splitting 
varible values occupied by each species. If FALSE, use min and max values occupied by each species.}
}
\description{
Make prediction dataframe for chopsticks
}
