% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-vocc-interaction.R
\name{plot_fuzzy_chopsticks}
\alias{plot_fuzzy_chopsticks}
\title{Plot interactions with confidence intervals}
\usage{
plot_fuzzy_chopsticks(
  model,
  y_label = "Predicted biomass trend",
  x_variable = "temp_trend_scaled",
  type = NULL,
  colours = NULL,
  species = NULL
)
}
\arguments{
\item{model}{TMB model with X_pj and pred_dat elements.}

\item{y_label}{Label response variable.}

\item{x_variable}{Variable to plot on x axis.}

\item{type}{Interaction type, if more than one in model.}

\item{colours}{Default NULL gives red for high and blue for low.}

\item{species}{Can plot just one species, otherwise NULL will facet.}
}
\description{
Plot interactions with confidence intervals
}
\examples{
plot_fuzzy_chopsticks(model, 
   y_label = "Predicted biomass trend", 
   x_variable = "temp_trend_scaled", 
   type = "temp")

}
