% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_vocc}
\alias{plot_vocc}
\title{Plot climate change vectors}
\usage{
plot_vocc(df, fill_col = NULL, fill_label = NULL, raster_alpha = 1,
  vec_aes = "distance", arrowhead_size = 0.015,
  vec_lwd_range = c(0.7, 0.8), vec_alpha = 1,
  max_vec_plotted = max(df$distance), min_vec_plotted = 4,
  NA_label = "NA", white_zero = min(fill, na.rm = TRUE) < 0,
  low_fill = "Steel Blue 4", mid_fill = "white", high_fill = "Red 3",
  vec_col = "white", coast = NULL, contours = NULL,
  axis_lables = FALSE, viridis_option = "D", viridis_dir = 1,
  transform_col = no_trans, raster_limits = NULL,
  legend_position = c(0.2, 0.25))
}
\arguments{
\item{df}{Dataframe produced by dist_based_vocc function.}

\item{fill_col}{Vector name for colouring raster grid cells.}

\item{fill_label}{Label for legend for raster grid cell colour.}

\item{raster_alpha}{Raster transparency.}

\item{vec_aes}{Vector name for any plotting variation.}

\item{arrowhead_size}{Changes head size for custom geom_quiver function.}

\item{vec_lwd_range}{Range in vector line widths.}

\item{vec_alpha}{Vector transparency.}

\item{max_vec_plotted}{Upper limit for vector lengths.
Cells with no match less than this distance will be labeled with [NA_label].}

\item{min_vec_plotted}{Lower limit for vector lengths actually plotted.}

\item{NA_label}{Symbol used to indicate cells with no analog. Default is "NA".}

\item{white_zero}{If TRUE, will always plot on custom fill scale.
Default will plot on this scale only if raster has negative values.}

\item{low_fill}{Colour of negative values if raster values span zero.}

\item{mid_fill}{Colour of zero value raster cells.}

\item{high_fill}{Colour of positive values if raster values span zero.}

\item{vec_col}{Colour of vectors.}

\item{coast}{Coast polygons where (x = "X", y = "Y", group = "PID").
If NULL, will attempt to create them for xy values in df.}

\item{contours}{Polygons of contour lines where (x = "X", y = "Y", group = "paste(PID, SID)").
If NULL, will attempt to create bathymetry layer for xy values in df using gfplot.}

\item{axis_lables}{Logical for inclusion of axis labels.}

\item{viridis_option}{Change between viridis colormap options available in ggplot.}

\item{viridis_dir}{Option to flip scale by giving value of -1.}

\item{transform_col}{Apply transformation to colour scale.
Accepts standard options (e.g. "sqrt") or unquoted custom transformations
defined using scales::trans_new (e.g. fourth_root_power).
Default is to apply no transformation (no_trans).}

\item{raster_limits}{Range of values to plot; those in excess will be red. Default of "NULL" plots full range.}

\item{legend_position}{Vector of coordinates for legend placement. Or "none" to remove legend.}
}
\description{
Plot climate change vectors
}
