% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys-spatial.R
\name{interp_survey_bathymetry}
\alias{interp_survey_bathymetry}
\alias{add_missing_depths}
\alias{scale_survey_predictors}
\title{Interpolate survey bathymetry}
\usage{
interp_survey_bathymetry(dat, utm_zone = 9)

add_missing_depths(dat, survey, years, bath)

scale_survey_predictors(dat)
}
\arguments{
\item{bath}{Output from [interp_survey_bathymetry()].}
}
\description{
Interpolate survey bathymetry

Add missing depths

Scale survey predictors
}
